Scale track text when making a vertical zoom
The text inside OrbitGl always had the same font size, so it was
impossible to scale track related text and not scale overlay text.

Now, we can have different font sizes and scale the text within the
track in the same way as other objects.

Related to b/162797051.

+++ OrbitCore/Params.h
@@ -5,8 +5,10 @@
 #ifndef ORBIT_CORE_PARAMS_H_
 #define ORBIT_CORE_PARAMS_H_
 
+#include <stdint.h>
+
 struct Params {
-  const float font_size = 14.f;
+  const uint32_t font_size = 14;
 };
 
 extern Params GParams;
